@@ -1,7 +1,7 @@
 [package]
 name = "tetromino-tiling-solver"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [lib]
 crate-type = ["cdylib", "rlib"] # wasm and native
